requirements.txt
matplotlib==3.7.2
Pillow==9.3.0
Pillow==10.0.0
Pillow==10.0.0
transformers==4.31.0
